+++ packages/integration-sdk-runtime/src/api/__tests__/index.test.ts
@@ -6,7 +6,9 @@ import {
   getApiKeyFromEnvironment,
   createApiClient,
   getAccountFromEnvironment,
+  compressRequest,
 } from '../index';
+import { AxiosRequestConfig } from 'axios';
 
 jest.mock('@lifeomic/alpha');
 
@@ -103,5 +105,41 @@ describe('createApiClient', () => {
         maxTimeout: 20000,
       },
     });
+  });
+});
+
+describe('compressRequest', () => {
+  it('should compress the request data when the URL matches', async () => {
+    const config: AxiosRequestConfig = {
+      method: 'post',
+      url: '/persister/synchronization/jobs/478d5718-69a7-4204-90b7-7d9f01de374f/entities',
+      headers: {},
+      data: { some: 'data' },
+    };
+
+    await compressRequest(config);
+
+    // Check if the 'Content-Encoding' header is set to 'gzip'
+    expect(config.headers!['Content-Encoding']).toBe('gzip');
+
+    // Check if the data is compressed
+    expect(config.data).toBeInstanceOf(Buffer);
+  });
+
+  it('should not compress the request when the URL does not match', async () => {
+    const config = {
+      method: 'post',
+      url: '/other-url',
+      headers: {},
+      data: { some: 'data' },
+    };
+
+    await compressRequest(config);
+
+    // Check that the 'Content-Encoding' header is not set
+    expect(config.headers['Content-Encoding']).toBeUndefined();
+
+    // Check that the data is not compressed
+    expect(config.data).toEqual({ some: 'data' });
   });
 });