+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to
 
 # Unreleased
 
+- Allow loading modules by name (ex. @jupiterone/graph-rumble) in the
+  `generate-ingestion-sources-config` and `generate-integration-graph-schema`
+  commands
+
 ## 11.0.3
 
 - Moved `vis` dependency to devDependencies