+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to
 
 # Unreleased
 
+- Add `compressUploads` to `CreateAPIClientOptions`. Enabling this will send
+  gzipped payloads.
+
 - Allow loading modules by name (ex. @jupiterone/graph-rumble) in the
   `generate-ingestion-sources-config` and `generate-integration-graph-schema`
   commands